# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/menus": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "查询数据",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "分页索引",
                        "name": "current",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "分页大小",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "查询值",
                        "name": "queryValue",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态(1:启用 2:禁用)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "显示状态(1:显示 2:隐藏)",
                        "name": "showStatus",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "父级ID",
                        "name": "parentID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/schema.ListResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "list": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/schema.Menu"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "创建数据",
                "parameters": [
                    {
                        "description": "创建数据",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.Menu"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.IDResult"
                        }
                    },
                    "400": {
                        "description": "{error:{code:0,message:无效的请求参数}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            }
        },
        "/api/v1/menus.tree": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "查询菜单树",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "状态(1:启用 2:禁用)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "父级ID",
                        "name": "parentID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/schema.ListResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "list": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/schema.MenuTree"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            }
        },
        "/api/v1/menus/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "查询指定数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "唯一标识",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Menu"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "404": {
                        "description": "{error:{code:0,message:资源不存在}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "更新数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "唯一标识",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新数据",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.Menu"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status:OK}",
                        "schema": {
                            "$ref": "#/definitions/schema.StatusResult"
                        }
                    },
                    "400": {
                        "description": "{error:{code:0,message:无效的请求参数}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "删除数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "唯一标识",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status:OK}",
                        "schema": {
                            "$ref": "#/definitions/schema.StatusResult"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            }
        },
        "/api/v1/menus/{id}/disable": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "禁用数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "唯一标识",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status:OK}",
                        "schema": {
                            "$ref": "#/definitions/schema.StatusResult"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            }
        },
        "/api/v1/menus/{id}/enable": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "启用数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "唯一标识",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status:OK}",
                        "schema": {
                            "$ref": "#/definitions/schema.StatusResult"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schema.ErrorItem": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "type": "integer"
                },
                "message": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        },
        "schema.ErrorResult": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "错误项",
                    "$ref": "#/definitions/schema.ErrorItem"
                }
            }
        },
        "schema.IDResult": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "schema.ListResult": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "object"
                },
                "pagination": {
                    "$ref": "#/definitions/schema.PaginationResult"
                }
            }
        },
...
        "schema.Menu": {
            "type": "object",
            "required": [
                "name",
                "show_status",
                "status"
            ],
            "properties": {
                "actions": {
                    "description": "动作列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.MenuAction"
                    }
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "creator": {
                    "description": "创建者",
                    "type": "string"
                },
                "icon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "string"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "parent_id": {
                    "description": "父级ID",
                    "type": "string"
                },
                "parent_path": {
                    "description": "父级路径",
                    "type": "string"
                },
                "router": {
                    "description": "访问路由",
                    "type": "string"
                },
                "sequence": {
                    "description": "排序值",
                    "type": "integer"
                },
                "show_status": {
                    "description": "显示状态(1:显示 2:隐藏)",
                    "type": "integer"
                },
                "status": {
                    "description": "状态(1:启用 2:禁用)",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
        "schema.MenuAction": {
            "type": "object",
            "required": [
                "code",
                "menu_id",
                "name"
            ],
            "properties": {
                "code": {
                    "description": "动作编号",
                    "type": "string"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "string"
                },
                "menu_id": {
                    "description": "菜单ID",
                    "type": "string"
                },
                "name": {
                    "description": "动作名称",
                    "type": "string"
                },
                "resources": {
                    "description": "资源列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.MenuActionResource"
                    }
                }
            }
        },
        "schema.MenuActionResource": {
            "type": "object",
            "required": [
                "method",
                "path"
            ],
            "properties": {
                "action_id": {
                    "description": "菜单动作ID",
                    "type": "string"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "string"
                },
                "method": {
                    "description": "资源请求方式(支持正则)",
                    "type": "string"
                },
                "path": {
                    "description": "资源请求路径（支持/:id匹配）",
                    "type": "string"
                }
            }
        },
        "schema.MenuTree": {
            "type": "object",
            "properties": {
                "actions": {
                    "description": "动作列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.MenuAction"
                    }
                },
                "children": {
                    "description": "子级树",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.MenuTree"
                    }
                },
                "icon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "string"
                },
                "name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "parent_id": {
                    "description": "父级ID",
                    "type": "string"
                },
                "parent_path": {
                    "description": "父级路径",
                    "type": "string"
                },
                "router": {
                    "description": "访问路由",
                    "type": "string"
                },
                "sequence": {
                    "description": "排序值",
                    "type": "integer"
                },
                "show_status": {
                    "description": "显示状态(1:显示 2:隐藏)",
                    "type": "integer"
                },
                "status": {
                    "description": "状态(1:启用 2:禁用)",
                    "type": "integer"
                }
            }
        },
        "schema.PaginationResult": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "schema.StatusResult": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "状态(OK)",
                    "type": "string"
                }
            }
        },
...
    }
}